others_requirements.txt
numpy
pandas
pillow
pytorch-fid
kgcPy
jupyter

scipy==1.15.2
scikit-learn==1.6.1
scikit-image==0.25.2

onnx==1.17.0
onnxruntime==1.20.1

tqdm
torchsummary
torchinfo
matplotlib==3.10.0
seaborn==0.13.2
tensorboard==2.19.0







